pyproject.toml
[project]
name = "ai-assessment"
version = "0.1.0"
description = "AI systems assessment covering RAG, diagnosis, classification, and systems design"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.5",
    "numpy>=1.26",
    "openai>=1.40",
    "socksio>=1.0.0",
    "pymupdf>=1.24.9",
    "python-dotenv>=1.0",
    "reportlab>=4.2",
    "scikit-learn>=1.5",
    "sentence-transformers>=3.0",
    "torch>=2.3",
    "transformers>=4.44",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["section2", "section3"]

[tool.pytest.ini_options]
testpaths = ["tests", "section3"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"
